+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "kekstagram",
   "version": "26.0.0",
   "private": true,
@@ -11,6 +12,7 @@
     "url": "https://github.com/htmlacademy-univer-javascript-1/1861897-kekstagram-1/issues"
   },
   "devDependencies": {
+    "@faker-js/faker": "^7.5.0",
     "browser-sync": "2.27.10",
     "eslint": "8.15.0",
     "eslint-config-htmlacademy": "5.1.2"
@@ -23,8 +25,5 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  },
-  "dependencies": {
-    "@faker-js/faker": "^7.5.0"
   }
 }
